requirements.txt
Flask>=2.2.0
Flask-SQLAlchemy
Flask-Login
youtube-transcript-api>=0.6.0
werkzeug
trafilatura
flask
flask-login
flask-sqlalchemy
mistralai
psycopg2-binary
python-dotenv
werkzeug
youtube-transcript-api
flask
flask-login
flask-sqlalchemy
mistralai
psycopg2-binary
python-dotenv
requests
trafilatura
werkzeug
youtube-transcript-api
sqlalchemy